refactor: use TaskTrellisService in completeTaskTool
